Add UserToken functionality
A UserToken functionality has been implemented. This update includes a configuration for UserToken parameters in the appsettings.json file. A new UserTokenOptions class was included to handle these configurations.
Moreover, a UserTokenService is in place which provides a function to generate a user token. The services were also registered in the Program.cs file.