Update launchSettings for simplified profile configuration
The changes in the launchSettings.json file helped streamline the local development setup. We stripped redundant IIS settings and configuration for the http and IIS Express profiles. Besides, we also adjusted the https profile to disable the automatic launch of the browser and updated the application URLs. These modifications are meant to provide a cleaner, more manageable setup for local development, reducing complexity and potential sources of configuration errors.