Add email verification functionality to user registration
Added functionality for sending email verification after a new user registers. An HTML template for email verification is included and a MailTemplateService to load and parse these templates with user-specific data. The user service is extended with a method for sending the verification email using the above services, triggered after successful user registration. The JSON property "MailTemplates" has been added to the configuration to specify the location of email templates.