Added update methods to Repository classes
In the Repository classes, both individual and bulk update methods were added to enable updating of entities in the database. These methods were also added to the IRepository interface for code consistency and enforced adherence. The aim of this change is to expand the functionality of the repository classes and streamline the process of updating entities in the database.