Add default claims to new users
Added functionality to assign a set of default claims to a new user upon registration.

This update extends the ClaimService with methods to locate claims by their names, and also extends the UserClaimService with a method to fetch these default claims. When a new user registers, the default claims are now automatically attached to that user by the UserService. This feature ensures that a base level of access rights is guaranteed for every registered user.