Add root admin setup
The root admin setup has been added to the post startup configuration to create or update a root admin user in the system. RootAdminOptions has been added and configured. The setup process includes checking if such a user exists, if not, it creates one and sends an activation email. If the user already exists, the setup process makes sure the user has all the available claims.