Change Query method's accessibility in Repository
Changed the accessibility of the Query method in the Repository class from protected to public. This is done to allow it to be used in the IRepository interface. This adjustment allows more flexibility in managing tracking behavior in different situations without having to directly interact with the Repository class or the DbContext.