Add user management functionality to UI
This commit introduces user management feature on the UI end. A new page 'UserManagement.razor' is added which allows managing users on a front-end table. Users can now search users using emails and manage claims right from the table. Changes also include the addition of an API client 'UserApiClient' for backend communication, which includes endpoints to fetch all users, find an individual user and to get count of users. `UserDto` class is updated to remove 'HasVerifiedEmail' property and add 'Id' property. A new controller 'UserController' is introduced in the Accounts project to provide endpoints required by the UI. UI models 'AccountClaim' and 'User' are also added to facilitate the feature. Finally, 'UserMapper' is updated to map from User entity to UserDto and vice versa. The dependency is added in 'Program.cs' in the UI project to use the 'IUserApi' interface with HttpClient.