Add claim management feature to User Panel
Implemented the claim management feature on the User Panel. This was done by creating two new interfaces 'IUserClaimApi' and 'IClaimApi' for Add and Remove claim functionality. A new dialog 'ClaimsDialog' was also added to handle assigning and removing of claims to the user. Additionally, to fetch paginated list of claims and claims count an update was made in 'ClaimController'. The User model was also updated to change Claims field type to List for dynamic updating. This will allow users to manage their claims more efficiently.