Add ControllerClaimProviderService and AuthorizationFilter
Added ControllerClaimProviderService in Shared Services to handle controller based claims. AuthorizationFilter was introduced to manage authorization process based on these claims. Claim mapping convention was also added for controllers. Now, authorization can be done on the basis of controller specific actions which provides more granular control over app authorization. Modified the csproj file to include reference to Microsoft.AspNetCore.App due to the usage of its components in the new service and filter. The Initialize method was introduced in ControllerClaimProviderService to initialize controller claims. Adjustments also made in Program.cs and SharedModuleExtensions.cs for integration of the new service and filter.