Add retrieval of all distinct claims in ControllerClaimProviderService
In both ControllerClaimProviderService and IControllerClaimProviderService, methods and properties were updated to include the functionality to retrieve all distinct claims. A List, _claimList was added to store these claims. The Initialize method now also clears _claimList and adds distinct controllerClaims to it. This was required to escalate claim handling capabilities and ease the retrieval of all distinct claims.