Refactor claims handling and remove static claims list
Removed the static Claims.cs file and adjusted claim-related functionalities. The changes were necessary to improve flexibility and maintainability. With static claim definition, adding a new feature or claim needed modifications in this list which is not practical for dynamic systems.

Instead of defining specific claims statically in code, now the system will recognize claims dynamically with the help of the new IControllerClaimProviderService. Service methods were modified and optimized for this behavior, and usage in various parts of the codebase was updated accordingly. This change significantly enhances system flexibility, making it easier to manage and extend in the future.