Update AccountClaim model and equality comparison
AccountClaim model has been restructured from a record to a class for better flexibility and compliance with project conventions. 'EqualityComparer' was removed and Equals method was overridden in the class itself to handle the equality comparison more efficiently. 'IsDefault' property was added to allow setting default account claim. Updates were reflected in the 'ClaimsDialog' UI component. It helps to more efficiently handle account claims and aligns with the overall project structure.