Add user claim management feature
This commit introduces user claim management feature.
In the navigation menu, a menu item for Users has been added under the Account group.
Moreover, a method for setting a claim as default was introduced in the IClaimService interface, implemented in the ClaimService class and added to the IClaimApi client interface.
A ClaimManagement razor page was added to the UI, which allows user to view claims and to toggle their default state.
For default state toggling, a corresponding endpoint was added to the ClaimController.
This feature will allow administrators to manage user claims more efficiently and will improve overall user management.