Refactor Cors setup to use extension methods
The changes in this commit move the Cors setup into an extension method which is added to the WebApplicationBuilder. This neatly encapsulates code into a reusable method and simplifies the original Cors setup in Program.cs.

Furthermore, CorsOptions has been relocated from the Accounts component to the Shared namespace to make it accessible across the project. These modifications contribute to better code organization and readability.