Add UserDataService and UserDataController for user data management
This commit introduces a UserDataService and UserDataController used to manage the user's data.

The UserDataService includes methods for finding, creating, and saving the user's data. It is registered as a transient service in the dependency injection container. A UserDataController is also implemented to expose these methods via HTTP Get and Post actions.

Moreover, a new project reference to the GameServicesCloud.Persistence.Abstractions has been added in the .csproj file. This was necessary because the IUserDataService interface has been moved to this referenced project.

This change should enable better user data tracking and management in the application.