Add UpdatedAt field to UserData model
A nullable DateTime field 'UpdatedAt' has been added to the UserData model for tracking the timestamp of the last update to the user record. The field has been added to the migrations, the model snapshot and the UserDataRepository. Now, every time the Save or Update method is called for user data, the 'UpdatedAt' field will be refreshed with the current DateTime. This makes it easier to track changes and updates to user data.