Add size constraint for UserData
The changes introduce a new UserDataOptions configuration setting that contains maximum byte size for UserData. This was implemented to apply a size limitation on the User data being saved into the system. Now, before saving the User data, its size is checked and if it exceeds the maximum size limit, an informative log is written and the data is not saved. This helps to prevent storage overflows for large data and ensures efficient use of storage.