Update Refit backend client setup to allow specifying backend name
This commit modifies the 'AddRefitBackendClient' method in both 'Program.cs' and 'RefitDiExtensions.cs' to include a new parameter, 'backendName'. Previously, the method hard-coded the use of the "Accounts" backend base URL, which was limiting as it only allowed interaction with one backend. With this change, the application increases flexibility by allowing different backends to be specified when setting up Refit backend clients, improving the scalability for more complex projects.