Add UserData admin controls and admin DTO
Implemented an UserController to access user data limited to admin access. A new UserDataAdminDto is created to limit the data displayed to an admin. This allows admins to oversee user data, and have control over the data without exposing too much information to the admin.

The UserController provides endpoints to retrieve all user data with pagination and filter capabilities, and count users based on a filter. The UserDataMapper was extended to map UserData to UserDataAdminDto.