Add Persistence service to UI
Added a persistence service configuration to appsettings and a corresponding client in Program.cs. The persistence service is used to manage user data, for which a new UserData model has been added to Models. The service will be accessed through the newly-created IUserDataApi interface in the Persistence folder. It supports finding all user data and counting user data with a filter.