Add user data deletion functionality
Added the ability to delete user data from the UI with a new endpoint and service method to delete data by user id in the UserDataApi and UserDataController respectively. To confirm the deletion, a confirmation dialog has been introduced on the UI. These changes were made to allow admins to remove user specific data when necessary according to user requests or system maintenance.