Add PaginationController for code reuse
Added a new abstract PaginationController that is now inherited by UserDataController, ClaimController and UserController which helps to maintain the DRY principle. This change helps to:
- Reuse pagination logic across controllers.
- Centralize the filtering logic.
IPaginator interface and implementation have been updated as well to support the use case where no filter is provided. Overall, these changes provide a more maintainable and efficient solution to handle entity listing and counting with optional filtering.