Add scores to LeaderboardDto
A new field has been added to the LeaderboardDto class to hold the count of scores. This is to provide the total number of scores on the leaderboard in the DTO model which is being mapped from the entity model. This modification will allow users to see the total number of scores on each leaderboard.