Implement dashboard access authorization policy
This commit incorporates a new authorization policy, "DashboardAccess", affecting the UI layouts. Firstly, the 'AuthorizeView' tags in 'MainLayout.razor' file are updated to use this policy to restrict access. Secondly, the authentication state is cached once on initialization to reduce duplication of code. The redundant 'GetUserName' function was removed as a result. An 'AuthorizedLayout.razor' file was added to restrict access to authenticated users fulfilling the new policy. This layout shows an error message and a logout button for unauthorized users. Lastly, the policy itself was defined in 'Program.cs' using a claim read from app settings. These changes tighten the security by restricting access to critical sections of the dashboard.